chore: migrate store extraction
